refactor(rdna-compute): split norm/activation methods into norm.rs
